+++ arangod/RocksDBEngine/RocksDBPrimaryIndex.cpp
@@ -157,7 +157,7 @@ RocksDBToken RocksDBPrimaryIndex::lookupKey(transaction::Methods* trx,
                           static_cast<uint32_t>(key.string().size()));
     if (f.found()) {
       rocksdb::Slice s(reinterpret_cast<char const*>(f.value()->value()),
-                       static_cast<size_t>(f.value()->valueSize));
+                       f.value()->valueSize());
       return RocksDBToken(RocksDBValue::revisionId(s));
     }
   }
